Python 3: Fix another integer/float bug

+++ tendenci/apps/events/forms.py
@@ -2283,7 +2283,7 @@ class StandardRegAdminForm(forms.Form):
 
 def add_months(sourcedate, months):
     month = sourcedate.month - 1 + months
-    year = sourcedate.year + month / 12
+    year = sourcedate.year + int(month / 12)
     month = month % 12 + 1
     day = min(sourcedate.day, calendar.monthrange(year,month)[1])
     return date(year,month,day)
